+++ CHANGELOG.md
@@ -5,6 +5,12 @@
 _(add items here for easier creation of next log entry)_
 
 - Add ability to specify a `defaultValue` to prefill the input.
+- When user has selected an option with the keyboard, blurring will select.
+- When user has no selected but autoselect is on, blurring will select.
+- Hovering no longer selects, just focuses.
+- When hovering out of component, focus returns to selected.
+- Allow enter to submit forms when menu isn't opened.
+- Hide results when going under minLength.
 
 ## 0.3.0 - 2017-03-09